fix elixir 1.4 compile warnings
