+++ lib/ex_admin/index.ex
@@ -21,7 +21,7 @@ defmodule ExAdmin.Index do
         register_resource MyProject.MyModel do
 
           index do
-            selectable_column
+            selectable_column()
 
             column :id
             column :name
