+++ lib/mix/tasks/admin.gen.resource.ex
@@ -32,14 +32,14 @@ defmodule Mix.Tasks.Admin.Gen.Resource do
     dest_path = Path.join ~w(web admin)
     dest_file_path = Path.join dest_path, filename
     source_file = Path.join([package_path | ~w(priv templates admin.gen.resource resource.exs)] )
-    source = source_file  |> EEx.eval_file(base: get_module, resource: module)
+    source = source_file |> EEx.eval_file(base: get_module(), resource: module)
     status_msg "creating", dest_file_path
     File.write! dest_file_path, source
     display_instructions config
   end
 
   defp display_instructions(config) do
-    base = get_module
+    base = get_module()
     IO.puts ""
     IO.puts "Remember to update your config file with the resource module"
     IO.puts ""
@@ -54,7 +54,7 @@ defmodule Mix.Tasks.Admin.Gen.Resource do
   end
 
   defp parse_args([module]) do
-    %Config{module: module, package_path: get_package_path}
+    %Config{module: module, package_path: get_package_path()}
   end
 
 end
