+++ test/integration/delete_test.exs
@@ -2,7 +2,7 @@ defmodule TestExAdmin.DeleteTest do
   use TestExAdmin.AcceptanceCase
   alias TestExAdmin.{Noid}
 
-  hound_session
+  hound_session()
   # Start hound session and destroy when tests are run
   setup do
     user = insert_user()
@@ -17,16 +17,16 @@ defmodule TestExAdmin.DeleteTest do
   test "delete a noid", %{user: user} do
     noid = insert_noid(user_id: user.id, name: "controller 1")
     _user2 = insert_user()
-    current_window_handle |> maximize_window
+    current_window_handle() |> maximize_window
     navigate_to admin_resource_path(Noid, :index)
 
-    assert page_source =~  noid.name
+    assert page_source() =~  noid.name
     click(find_element(:class, "delete_link"))
 
     assert dialog_text() == "Are you sure you want to delete this?"
     accept_dialog()
 
-    assert current_url ==  admin_resource_path(Noid, :index)
-    refute page_source =~ noid.name
+    assert current_url() ==  admin_resource_path(Noid, :index)
+    refute page_source() =~ noid.name
   end
 end
