+++ web/controllers/admin_controller.ex
@@ -22,7 +22,7 @@ defmodule ExAdmin.AdminController do
   def select_theme(conn, %{"id" => id} = params) do
     {id, _} = Integer.parse(id)
     {_, theme} = Application.get_env(:ex_admin, :theme_selector, []) |> Enum.at(id)
-    loc = Map.get(params, "loc", admin_path) |> URI.parse |> Map.get(:path)
+    loc = Map.get(params, "loc", admin_path()) |> URI.parse |> Map.get(:path)
 
     Application.put_env :ex_admin, :theme, theme
     redirect conn, to: loc
@@ -34,7 +34,7 @@ defmodule ExAdmin.AdminController do
     repo = Application.get_env(:ex_admin, :repo)
     user = repo.get current_user.__struct__, id
     require Logger
-    loc = Map.get(params, "loc", admin_path) |> URI.parse |> Map.get(:path)
+    loc = Map.get(params, "loc", admin_path()) |> URI.parse |> Map.get(:path)
     {mod, fun} = Application.get_env :ex_admin, :logout_user
     apply mod, fun, [conn]
     conn = conn |> assign(:current_user, user)
